use colors.primary everywhere
Prior to this commit we had a large number of colors defined at the element level, meaning changing your primary color in the tailwind config would not update it everywhere.

After this commit we're using the pattern of assigning the class "text-primary-weight" everywhere we want that primary color.

I removed the font weight definitions since we're being specific about the weight at each element anyway, so it mostly wasn't used. I would understand reverting that if we wanted to keep it for educational reasons.